+++ CHANGELOG.md
@@ -6,6 +6,7 @@ v1.1.2 (trunk)
 
 Bug Fixes:
  * Don't crash exporting to cscope if tokens overlap.
+ * In golang, don't parse inside string literals.
 
 v1.1.1 (2014-07-21)
 --------------------